Remove unreachable serialization code (#240)
There is code present in the library for container serialization
that is not reachable through the API.  This code is tied to the
`container_serialize()` and `container_deserialize()` functions,
which dispatch to functions for each specific container.  The
active serialization code paths are based on _write() and _read()
dispatch functions.

The only current usage of these functions was in a test of the `run`
container serialization in toplevel_unit.c.  This commit changes
that test so that it still generates a lower-level run container,
but then injects it into a single-container roaring_bitmap_t.  Then
it uses the active serialize and deserialize functions on that
bitmap in the test--effectively making it test deployed API code.

With that change in place, the following routines can be removed
(or put under an #ifdef if they are for some future purpose).

* ra_size_in_bytes()
    * container_serialization_len()
        * bitset_container_serialization_len()
        * array_container_serialization_len()
        * run_container_serialization_len()
* container_serialize()
    * bitset_container_serialize()
    * array_container_serialize()
    * run_container_serialize()
* container_deserialize()
    * bitset_cotainer_deserialize()
    * array_container_deserialize()
    * run_container_deserialize()

This commit removes them to demonstrate no loss of functionality.