Inplace andnot/xor/lazy_xor Shared container leak (#258)
Address sanitizer was reporting memory leaks when running the
`toplevel_unit` tests.  The leaks originated in the inplace
andnot/xor/lazy_xor code, and are connected to the choice to
not perform an inplace operation when a container is shared...
but to call into the non-modifying forms.

The and/or operations have code which checks if the resulting
container from an operation does not match the input and frees
the input container, like:

    if (c != c1) {  // in this instance a new container was
                    // created, and we need to free the old one
         container_free(c1, typecode1);
    }

There is no such code for andnot/xor/lazy_xor.  Instead the
inplace branches seem to do their own freeing of replaced
containers.  But the non-inplace branch does not account for the
needed drop of a refcount on the shared container that is being
replaced.

This adds a call to `shared_container_free()` on those branches,
which appears to stop the leaks in `toplevel_unit`.