Safer downcasts w/macros that static_cast<> in C++
C's sole cast operator will cast nearly anything to anything else, which
can be dangerous...and the container subclasses in roaring make heavy
use of downcasting from container types.  This is risky, and can cause
problems the compiler won't catch e.g. when the wrong level of indirection
is used:

    void some_func(container_t **c, ...) {  // double pointer, not single
        array_container_t *ac1 = (array_container_t *)(c);  // uncaught!!
    }

When compiling as C there's nothing that can be done about that.  However,
with `container_t` being an empty-base-class-optimization base of all the
container structs, the C++ build can do better by using static_cast<>.

    void some_func(container_t **c, ...) {  // double pointer, not single
        array_container_t *ac2 = static_cast<array_container_t *>(c);
             // ^-- compiler errors due to wrong level of indirection

        array_container_t *ac3 = static_cast<array_container_t *>(*c);
             // ^-- not an error due to legal downcasting pattern
    }

To take advantage of this while still building as C, a CAST() macro is
defined in the internal headers.  When building as C++ it will act as a
static cast, but uses old-style C casts in C mode.

As a further improvement in rolling this feature out to callsites, shortcut
macros for casting to container subclasses are provided:

     array_container_t *ac4 = CAST_array(c);
     const array_container_t *ac4 = const_CAST_array(c);

Having an uppercase portion to the name helps hint at the "not-a-function"
nature of the macro, while keeping the other parts lowercase help smooth
the visual impact at the callsites.  In addition to providing added
compile-time checking, the resulting code is shorter and cleaner.