hide containers.h/roaring_array.h from public api (#259)
* Fix for roaring_overwrite() when source is empty

ra_overwrite() lacked a protection present in ra_copy() which checks
to make sure that the input size is not 0.  If it is, it might be
trying to copy from a null pointer...which is not legal even when
size is 0:

https://www.imperialviolet.org/2016/06/26/nonnull.html

* Unify ra_copy() and ra_overwrite()

Finding a problem in ra_overwrite() that had been worked around in
ra_copy() drew attention to how similar the two routines were (and
that a difference was only a bug).

The semantics were that ra_copy() assumed it was writing into
uninitialized memory while ra_overwrite() was willing to reuse the
existing capacity of a roaring_array_t if it was big enough.

This reframes the roaring_bitmap_copy() routine as initializing a
list of containers of the appropriate size (so it won't need a
realloc_array()) and then calling ra_overwrite().  ra_copy() is
then deleted.

* hide containers.h/roaring_array.h from public api

The function roaring_bitmap_contains() in the C API was declared as
inline, pulling in a considerable set of dependencies.  Usage of the
container_contains() function brought in definitions for
bitset_container_t, array_container_t, run_container_t, as well as
all the needed support functions.

While inlining can give considerable advantages, the definition of
this particular public export as inline causes collateral damage
that is unlikely to move the needle much to client performance.

This migrates the implementation of roaring_bitmap_contains() into
roaring.c, and moves the `roaring_array_t` from roaring_array.h
into roaring_types.h.  This then allows containers.h to be removed
from the public API.

Code such as tests that wants to use the internal functions thus
must include the internal headers to do so.  (Using the internal API
to reuse code is also an axis that some clients might take in
pursuit of greater performance, if they were willing to deal with
internals changing.)

* Remove dependence of C++ class on roaring_array.h

This takes another step to reducing the export surface for the API,
by adding a missing routine needed by C++ to initialize a roaring
bitmap structure in client-controlled memory.

Previously the C++ class had to include roaring_array.h and use
ra_init(), otherwise the only way it could create a bitmap would
be through a malloc()'d handle.  roaring_bitmap_init_with_capacity()
is added to the API to provide this.

(roaring_bitmap_init_cleared() is created as an inline that calls
roaring_bitmap_init_capacity() with 0, to show that by contract in
the API that these are not semantically different.  Since it cuts
down on variation, the change is applied to roaring_bitmap_create()
as well, as an inline call to roaring_bitmap_create_with_capacity())

The other features used by the C++ class were ra_copy(), which
already had an export suitable for the functionality in
roaring_bitmap_overwrite() (rebased onto a bugfix for that). Also
ra_clear(), effectively available as roaring_bitmap_clear()

With this change, roaring_array.h and containers.h can be left out
of the usage of the C++ API as well; leading to *many* fewer exports.

* Explicitly #include memcpy in roaring64map.hh

The push toward minimal inclusion exposed unmentioned dependencies
in roaring64map.hh, which had previously implicitly gotten a define
for malloc() by including roaring.h but now does not.

This adds the include, and changes to use the scoped std::malloc()
notation required for when `<cstring>` is included vs string.h

* Use private/public separation in amalgamation.sh

Though the export dependencies for roaring.h and roaring.hh were
reduced, the amalgamation script put all the header files into
the aggregated roaring.h -- whether they were public or private.

This separates the headers out, so that the private ones are
packed into the .C file instead of the .H file.

The .HH file is also adjusted to use the #define guard that keeps
the C API from being put in the global namespace by default when
it re-adds the `#include roaring.h` that it stripped out.

Minor changes to reduce redundancy are included--factoring the
timestamp and generated file warning into a function, an omitting
the original-file-location comments when including the license.

* Some light tidying of amalgamation.sh

* Moves generated file list to top of file
* Roughly honors 80-character margin in script source
* Limits running status commentary to file being processed
* Meaningful newline breaks in script output
* More comments