movable_CAST for container_t** -> xx_container_t**
The CAST() macro when used in C++ was adding an extra check
using static_cast<> to make sure that container pointers were
correctly downcast, vs. the unsafe C cast that would turn
anything-into-anything.

Unfortunately, that doesn't work for pointer-to-pointer,
even though the conversion should be legal.  reinterpret_cast<>
works, but loses the benefit of the check for matching types
and levels of dereference.

This uses <type_traits> and a template to add a compile-time
static_assert<> that ensures correct conversion.