void* -> container_t*, checked in C++ builds (#261)
There are currently 4 subclasses of container in roaring.  As it is
a C codebase, their structs do not share a common "base class".  So
when passing around a generic container, void* had to be used.

This commit introduces a new type to capture the intention of a
generic container, called `container_t`.  It is defined as `void` in
the C build, and serves as documentation...which is in itself an
improvement.

However, the C++ build can define container_t as an actual base
class for the containers.  It can do so in a way that doesn't disrupt
the C compatible behavior of the structs, due to the derived class
being standard layout because of the "Empty Base Class Optimization"

https://en.cppreference.com/w/cpp/language/ebo

When this is used, you can assign any subclass xxx_container_t to a
container_t* without a cast...but unlike a void*, ONLY those
subclasses can be assigned this way.  Other type checking mechanisms
can be used as well that exploit the relationship.

In the spirit of improving namespacing and avoiding putting things
in global scope, this limits the definition of `container_t` to the
internal code.  But since container pointers are part of the public
definition of `roaring_array_t`, there is an alias for the base
class as a macro ROARING_CONTAINER_T.  It is a macro so it can be
undefined in the internal code to force use there of the shorter
name of `container_t`.

While the commit wasn't meant to be a reformatting crusade, the
`container_t` name is longer than `void` and caused disruption at
callsites...many of which were already hard to read.  And the number
of places that `void*` were used were very numerous.  In the interest
of making the code legible, the usage sites are adapted with some
shorter names being used (e.g. type1 vs. container_type_1).